changed link adress in home.js. renamed savedbooks to player.js and searchbooks.js to home.js
