+++ client/src/pages/Player.js
